Struct chip8_rs::chip8::Chip8

pub struct Chip8 {

    registers: [u8; 16],
    memory: [u8; 4096],
    index_register: u16,
    program_counter: u16,
    stack: [u16; 16],
    stack_pointer: u8,
    delay_timer: u8,
    sound_timer: u8,
    keypad: [u8; 16],
    pub video: [u32; 2048],
    op_code: u16,
    table: [fn(_: &mut Chip8); 16],
    table0: [fn(_: &mut Chip8); 15],
    table8: [fn(_: &mut Chip8); 15],
    tableE: [fn(_: &mut Chip8); 15],
    tableF: [fn(_: &mut Chip8); 102],
    debug_mode: bool,

}

General chip 8 struct

Fields
registers: [u8; 16]memory: [u8; 4096]index_register: u16program_counter: u16stack: [u16; 16]stack_pointer: u8delay_timer: u8sound_timer: u8keypad: [u8; 16]video: [u32; 2048]op_code: u16table: [fn(_: &mut Chip8); 16]table0: [fn(_: &mut Chip8); 15]table8: [fn(_: &mut Chip8); 15]tableE: [fn(_: &mut Chip8); 15]tableF: [fn(_: &mut Chip8); 102]debug_mode: bool
Implementations

impl Chip8

pub fn new() -> Self

Create a new chip8 instance with an empty rom, ready for use

Values

registers: all zeroes

memory/rom: all zeroes

index_register: 0

program_counter: 0x200

stack: all zeroes

stack_poionter: 0

delay_timer: 0

sound_timer: 0

keypad: all zeroes

video: all zeroes

op_code: 0

fontset_size: 80

pub fn add_table(&mut self)

Add the correct function pointer tables to the newly created Chip8 object

pub fn debug(&mut self)

Toggle debug mode for current chip8 instance

pub fn load_rom(&mut self, path: PathBuf)

Loads a given rom into memory, starting from memory address 0x200

pub fn dump_rom(&self, start_address: usize, byte_number: usize) -> Vec<u8>

Returns byte_number amount of bytes after the pointer in rom

pub fn dump_video(&mut self) -> Vec<u32>

pub fn pretty_print_video(&mut self)

fn load_fonts(&mut self)

Load the fontset into memory

pub fn rand_byte(&self) -> u8

Generate a random byte(0, 255)

pub fn OP_00E0(&mut self)

OPCODE 00E0 - Clear Screen

fn OP_00EE(&mut self)

OPCODE 00EE - Return from subroutine

fn OP_1nnn(&mut self)

OPCODE 1NNN - Jump to location NNN(set program counter to nnn)

fn OP_2nnn(&mut self)

OPCODE 2NNN - Call subroutine at location NNN

fn OP_3xkk(&mut self)

OPCODE 3XKK - Skip next instruction if Vx = kk Since our PC has already been incremented by 2 in Cycle(), we can just increment by 2 again to skip the next instruction.

fn OP_4xkk(&mut self)

OPCODE 4XKK - Skip next instruction if Vx != kk

fn OP_5xy0(&mut self)

OPCODE 5XY0 - Skip next instruction if Vx = Vy.

fn OP_6xkk(&mut self)

OPCODE 6XKK - Set Vx = kk.

fn OP_7xkk(&mut self)

OPCODE 7XKK - Set Vx = Vx + kk.

fn OP_8xy0(&mut self)

OPCODE 8XY0 - Set Vx = Vy.

fn OP_8xy1(&mut self)

OPCODE 8XY1 - Set Vx = Vx OR Vy.

fn OP_8xy2(&mut self)

OPCODE 8XY2 - Set Vx = Vx AND Vy

fn OP_8xy3(&mut self)

OPCODE 8XY3 - Set Vx = Vx XOR Vy

fn OP_8xy4(&mut self)

OPCODE 8XY4 - Set Vx = Vx + Vy, set VF = carry. The values of Vx and Vy are added together. If the result is greater than 8 bits (i.e., > 255,) VF is set to 1, otherwise 0. Only the lowest 8 bits of the result are kept, and stored in Vx.

fn OP_8xy5(&mut self)

OPCODE 8XY5 - Set Vx = Vx - Vy, set VF = NOT borrow. If Vx > Vy, then VF is set to 1, otherwise 0. Then Vy is subtracted from Vx, and the results stored in Vx.

fn OP_8xy6(&mut self)

OPCODE 8XY6 - Set Vx = Vx SHR 1. If the least-significant bit of Vx is 1, then VF is set to 1, otherwise 0. Then Vx is divided by 2.

fn OP_8xy7(&mut self)

OPCODE 8XY7 - SUBN Vx, Vy If Vy > Vx, then VF is set to 1, otherwise 0. Then Vx is subtracted from Vy, and the results stored in Vx.

fn OP_8xyE(&mut self)

OPCODE 8XYE - Set Vx = Vx SHL 1. If the most-significant bit of Vx is 1, then VF is set to 1, otherwise to 0. Then Vx is multiplied by 2.

fn OP_9xy0(&mut self)

OPCODE 9XY0 - Skip next instruction if Vx != Vy

fn OP_Annn(&mut self)

OPCODE ANNN - set I = nnn

fn OP_Bnnn(&mut self)

OPCODE BNNN - Jump to location nnn + V0

fn OP_Cxkk(&mut self)

OPCODE CXKK - Set Vx = random byte AND kk.

fn OP_Dxyn(&mut self)

OPCODE DXYN - Display n-byte sprite starting at memory location I at (Vx, Vy), set VF = collision.

fn OP_Ex9E(&mut self)

OPCODE EX9E - Skip next instruction if key with the value of Vx is pressed.

fn OP_ExA1(&mut self)

OPCODE EXA1 - Skip next instruction if key with the value of Vx is not pressed

fn OP_Fx07(&mut self)

OPCODE FX07 - Set Vx = delay timer value

fn OP_Fx0A(&mut self)

OPCODE FX0A - Wait for a key press, store the value of the key in Vx.

fn OP_Fx15(&mut self)

OPCODE FX15 - Set delay timer = Vx.

fn OP_Fx18(&mut self)

OPCODE FX18 - Set sound timer = Vx.

fn OP_Fx1E(&mut self)

OPCODE FX1E - Set I = I + Vx.

fn OP_Fx29(&mut self)

OPCODE FX29 - Set I = location of sprite for digit Vx.

fn OP_Fx33(&mut self)

OPCODE FX33 - Store BCD representation of Vx in memory locations I, I+1, and I+2. The interpreter takes the decimal value of Vx, and places the hundreds digit in memory at location in I, the tens digit at location I+1, and the ones digit at location I+2.

fn OP_Fx55(&mut self)

OPCODE FX55 – Store registers V0 to VX in memory starting at location X

fn OP_Fx65(&mut self)

OPCODE FX65 - Read registers V0 through Vx from memory starting at location I.

fn OP_ERR(&mut self)

Fallback OPCODE

fn Table0(&mut self)

fn Table8(&mut self)

fn TableE(&mut self)

fn TableF(&mut self)

pub fn Cycle(&mut self)